+++ src/components/motion/motion.test.tsx
@@ -13,7 +13,7 @@ describe("Motion", () => {
   it("should display raw image count when no tilt. align. is present", async () => {
     renderWithProviders(<Motion parentType='dataCollections' parentId={2} />);
 
-    await screen.findByText("10");
+    await expect(screen.findByLabelText("Total Pages")).resolves.toHaveTextContent("10");
   });
 
   it("should display enabled comments button when comments are present", async () => {
@@ -40,5 +40,17 @@ describe("Motion", () => {
     renderWithProviders(<Motion parentType='tomograms' onTotalChanged={totalChanged} parentId={3} />);
 
     await waitFor(() => expect(totalChanged).toBeCalled());
+  });
+
+  it("should respect current page prop if passed", async () => {
+    const totalChanged = jest.fn();
+    renderWithProviders(<Motion currentPage={5} parentType='tomograms' onTotalChanged={totalChanged} parentId={3} />);
+
+    expect(await screen.findByLabelText("Current Page")).toHaveValue("5");
+  });
+
+  it("displays '?' if passed values for raw total and total include NaN", async () => {
+    renderWithProviders(<Motion parentType='tomograms' parentId={4} />);
+    await screen.findByRole("heading", { name: "?" });
   });
 });