+++ src/components/motion/pagination.test.tsx
@@ -18,9 +18,15 @@ describe("MotionPagination", () => {
   });
 
   it("should respect default item set by user", async () => {
-    renderWithProviders(<MotionPagination displayDefault='90' total={112} />);
+    renderWithProviders(<MotionPagination displayDefault={90} total={112} />);
 
     expect(screen.getByDisplayValue("90")).toBeInTheDocument();
+  });
+
+  it("should not display default if default is an invalid page number", async () => {
+    renderWithProviders(<MotionPagination displayDefault={-1} total={112} />);
+
+    expect(screen.queryByDisplayValue("-1")).not.toBeInTheDocument();
   });
 
   it("should set middle item as default if startFrom is middle", async () => {