+++ src/components/visualisation/table.test.tsx
@@ -65,6 +65,6 @@ describe("GenericListing", () => {
     );
 
     fireEvent.click(screen.getByRole("cell", { name: /value4/i }));
-    expect(callbackMock).toBeCalledWith({ key1: "value4", key2: "value5", key3: "value6" });
+    expect(callbackMock).toBeCalledWith({ key1: "value4", key2: "value5", key3: "value6" }, 1);
   });
 });