+++ src/store/store.ts
@@ -1,10 +1,8 @@
 import { configureStore, ThunkAction, Action } from '@reduxjs/toolkit';
-import authReducer from '../features/authSlice';
 import uiReducer from "../features/uiSlice"
 
 export const store = configureStore({
   reducer: {
-    auth: authReducer,
     ui: uiReducer
   },
 });