+++ CHANGELOG.rst
@@ -3,16 +3,27 @@ Changelog
 ==========
 
 ++++++++++
-v0.3.0 (23/01/2023)
+v0.3.0 (--/01/2023)
 ++++++++++
 
 **Added**
 
 - Data collection statistics page for single particle analysis
+- Users can now search through collections (tomogram) in a data collection group
+
+**Changed**
+
+- Resets listing page when searching
+- Performs auth before page is loaded, avoids multiple "refreshes"
+
+++++++++++
+v0.2.1 (27/01/2023)
+++++++++++
 
 **Changed**
 
 - Every request displays loading bar, instead of loading bar state being changed by hand
+- Fixes tilt alignment axis display in tomograms
 
 ++++++++++
 v0.2.0 (23/01/2023)